android: Add flag for external camera support
In the static metadata, we claim that we don't support external cameras,
yet libcamera core defaults the camera location to external. In the HAL
CameraDevice we defaulted external to front, but we hadn't handled
CameraHalManager when counting the internal vs external cameras for
hal_get_number_of_cameras().

To solve this, add a flag to hold if we have external camera support
(hardcoded to false until we do), and add that to the condition in
CameraHalManager when deciding whether a camera should be counted as
internal or external. CameraDevice also uses this flag to decide on the
default location.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>