ipa: libipa: Provide a common base for FrameContexts
Provide a common IPAFrameContext as a base for IPA modules to inherit
from.

This will allow having a common set of parameters for every FrameContext.

We still have to use a templated instance of the FCQueue though to make
sure we allocate the correct types.

Both the RKISP1 and the IPU3 create IPA specific FrameContext structures
which inherit from the IPAFrameContext.

While adjusting interface to Algorithm::process() the FrameContext is
converted from a pointer to a reference.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>