fixup: Harden frame context overwrite error condition
Harden the error condition in the case the desired frame context has
been forever lost as too many requests has been queued by the pipeline
handler. As this is a programming error, it is fair to ASSERT() on it.

While at it expand the comment on the opposite case, where a request
underrun makes the queue overflow and record that we might possibily
want to notify that to applications.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>